finished folder structure
