+++ src/app/actions/mathActions.js
@@ -1,0 +1,13 @@
+export function addNumber(number) {
+    return {
+        type: "ADD",
+        payload: number
+    };
+}
+
+export function subtractNumber(number) {
+    return {
+        type: "SUBTRACT",
+        payload: number
+    };
+}
