+++ src/app/actions/userActions.js
@@ -1,0 +1,13 @@
+export function setName(name) {
+    return {
+        type: "SET_NAME",
+        payload: name
+    };
+}
+
+export function setAge(age) {
+    return {
+        type: "SET_AGE",
+        payload: age
+    };
+}
