+++ src/app/components/User.js
@@ -1,20 +1,18 @@
 import React from "react";
 
-export class User extends React.Component {
-    render() {
-        return (
-            <div>
-                <div className="row">
-                    <div className="col-xs-12">
-                        <h1>The User Page</h1>
-                    </div>
-                </div>
-                <div className="row">
-                    <div className="col-xs-12">
-                        <p>User Name: {this.props.username}</p>
-                    </div>
+export const User = (props) => {
+    return (
+        <div>
+            <div className="row">
+                <div className="col-xs-12">
+                    <h1>The User Page</h1>
                 </div>
             </div>
-        );
-    }
+            <div className="row">
+                <div className="col-xs-12">
+                    <p>User Name: {props.username}</p>
+                </div>
+            </div>
+        </div>
+    );
 }
