+++ src/app/containers/App.js
@@ -1,8 +1,9 @@
 import React from "react";
 import {connect} from "react-redux";
 
-import { User } from './User';
-import { Main } from './Main';
+import { User } from "../components/User";
+import { Main } from "../components/Main";
+import { setName } from "../actions/userActions";
 
 class App extends React.Component {
     render() {
@@ -25,10 +26,7 @@ const mapStateToProps = (state) => {
 const mapDispatchToProps = (dispatch) => {
     return {
         setName: (name) => {
-            dispatch({
-                type: "SET_NAME",
-                payload: name
-            });
+            dispatch(setName(name));
         }
     };
 };
