+++ src/app/store.js
@@ -1,0 +1,14 @@
+import {createStore, combineReducers, applyMiddleware} from "redux";
+import logger from "redux-logger";
+
+import math from "./reducers/mathReducer";
+import user from "./reducers/userReducer";
+
+export default createStore(
+    combineReducers({
+        math,
+        user
+    }),
+    {},
+    applyMiddleware(logger())
+);
